+++ libs/dh/market-participant/routing/src/index.ts
@@ -15,6 +15,6 @@
  * limitations under the License.
  */
 export * from './lib/dh-market-participant-path';
-export * from './lib/dh-market-participant-actor-path';
+export * from './lib/dh-market-participant-actors-path';
 export * from './lib/dh-market-participant-organizations-path';
 export * from './lib/dh-market-participant-gridarea-path';
